move ExtractNativePropsFromDefault to types folder
- doesn't make much sense from folder structure perspective
