+++ project-base/storefront/components/Forms/Lib/ChoiceFormLine.tsx
@@ -1,6 +1,6 @@
 import { twMergeCustom } from 'helpers/twMerge';
 import { HTMLAttributes } from 'react';
-import { ExtractNativePropsFromDefault } from 'typeHelpers/ExtractNativePropsFromDefault';
+import { ExtractNativePropsFromDefault } from 'types/ExtractNativePropsFromDefault';
 
 type NativeProps = ExtractNativePropsFromDefault<HTMLAttributes<HTMLDivElement>, never, 'style'>;
