+++ project-base/storefront/components/Forms/Radiobutton/Radiobutton.tsx
@@ -2,7 +2,7 @@ import { LabelWrapper } from '../Lib/LabelWrapper';
 import { Image } from 'components/Basic/Image/Image';
 import { ImageSizesFragmentApi } from 'graphql/generated';
 import { forwardRef, InputHTMLAttributes, MouseEventHandler, ReactNode, useCallback } from 'react';
-import { ExtractNativePropsFromDefault } from 'typeHelpers/ExtractNativePropsFromDefault';
+import { ExtractNativePropsFromDefault } from 'types/ExtractNativePropsFromDefault';
 
 type NativeProps = ExtractNativePropsFromDefault<
     InputHTMLAttributes<HTMLInputElement>,
