+++ project-base/storefront/types/ExtractNativePropsFromDefault.ts
